Device Id Generator
I have created this device ID generator for tracking user journeys in iOS apps. This device ID will only be deleted if the data is deleted from the cloud. Deleting the app or clearing the cache will not remove this device ID. Also this device id generator work for all platform.