SSL Pinning
SSL pinning is a security measure that allows you to ensure that your app is communicating with the intended server. It works by pinning the server's public certificate to your app. When your app makes a connection to the server, it verifies the server's certificate against the pinned certificate. If the certificates do not match, the connection is aborted.

SSL pinning can help to protect your app from man-in-the-middle attacks. In a man-in-the-middle attack, an attacker intercepts the communication between your app and the server and impersonates the server. By pinning the server's certificate, you can prevent your app from being fooled by the attacker.

To implement SSL pinning in iOS Swift, you can use the TrustKit library. TrustKit is a third-party library that provides a simple and easy-to-use way to implement SSL pinning.

Here is an example of how to implement SSL pinning in iOS Swift using TrustKit: